feat: Select a matching version of the Ansible playbooks by default
Without this change, the version to use when cloning the Ansible
playbooks is the repository's main branch.

This means that an image file built for a specific version of this
plugin may differ in functionality, based on the state of the playbook
repo's main branch at build time.

Thus, change the default version to be cloned to match that of this
plugin. For example, for the 2.1.0 version of this plugin, clone the
"v2.1.0" tag of the playbook branch by default. This can be overridden
by setting ENROLLMENTREPORTS_REVISION in config.yml, just as before.

This means that whenever we make a change to the plugin *or* the
playbooks, we need to bump the tag in both, which ensures that
they always match correctly.